+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Pimple Service Provider
+
 ## 0.1.0 - 2017-05-15
 
 Initial Release